exported the traits implementation to the separate impls module as per the review comment
