+++ pallets/disputes/src/lib.rs
@@ -4,6 +4,7 @@
 // avaliable methods etc.
 
 pub use pallet::*;
+pub mod impls;
 pub mod traits;
 use core::fmt::Debug;
 use frame_support::{pallet_prelude::*, weights::Weight};
@@ -13,8 +14,8 @@ pub mod pallet {
     use super::*;
     use codec::{FullCodec, FullEncode};
     use frame_system::pallet_prelude::*;
-    use sp_runtime::{traits::AtLeast32BitUnsigned, DispatchError};
-    use traits::{DisputeRaiser,DisputeHooks};
+    use sp_runtime::traits::AtLeast32BitUnsigned;
+    use traits::DisputeHooks;
 
     #[pallet::pallet]
     pub struct Pallet<T>(_);
@@ -105,7 +106,7 @@ pub mod pallet {
                     // FELIX REVIEW: 
                     // This doesnt work, notice how a vote isnt mutable, therefore you arnt mutating it at all.
                     // Also how do you account for changing of a vote 
-                    let vote = &d.votes;
+                    let mut vote = &d.votes;
                     if is_yay {
                         vote.yay.saturating_add(1);
                     } else {
@@ -138,107 +139,27 @@ pub mod pallet {
     }
 
 
-    // FELIX REVIEW: 
-    // Add all the impl<x: y>... into its own folder called impls.rs
-    impl<T: Config> DisputeRaiser<AccountIdOf<T>> for Pallet<T> {
-        type DisputeKey = T::DisputeKey;
-        type MaxReasonLength = <T as Config>::MaxReasonLength;
-        type MaxJurySize = <T as Config>::MaxJurySize;
-
-        fn raise_dispute(
-            dispute_key: Self::DisputeKey,
-            raised_by: AccountIdOf<T>,
-            fund_account: AccountIdOf<T>,
-            reason: BoundedVec<u8, Self::MaxReasonLength>,
-            jury: BoundedVec<AccountIdOf<T>, Self::MaxJurySize>,
-        ) -> Result<(), DispatchError> {
-            // creating the struct with the passed information and initializing vote as 0 initially
-            let dispute: Dispute<T> = Dispute {
-                raised_by: raised_by.clone(),
-                fund_account,
-                votes: Vote { yay: 0, nay: 0 },
-                reason,
-                jury,
-            };
-
-            ensure!(
-                !Disputes::<T>::contains_key(dispute_key.clone()),
-                Error::<T>::DisputeAlreadyExists
-            );
-
-            //storing the raised dispute inside the disputes storage
-            Disputes::<T>::insert(dispute_key, dispute);
-
-            // Raise Event
-            //SHANKAR if want to add more information while raising a dispute like returning the whole dispute struct to
-            //get some more info about what has been raised or so on?
-            //FELIX REVIEW Only add information to events that you are certain youre gonna use on the app.
-            //each event is stored on chain so it would bloat the chain if we started returning whole structs.
-            Self::deposit_event(Event::DisputeRaised { who: raised_by });
-
-            Ok(())
-        }
-    }
-
+  
+   
     #[derive(Encode, Decode, PartialEq, Eq, Clone, Debug, TypeInfo, MaxEncodedLen)]
     #[scale_info(skip_type_params(T))]
     pub struct Dispute<T: Config> {
-        raised_by: AccountIdOf<T>,
-        fund_account: AccountIdOf<T>,
+        pub raised_by: AccountIdOf<T>,
+        pub fund_account: AccountIdOf<T>,
         // TODO: Add balance type
         // currencyid: CurrencyId
         //fund_amount: BalanceOf<T>
-        votes: Vote,
-        reason: BoundedVec<u8, <T as Config>::MaxReasonLength>,
-        jury: BoundedVec<AccountIdOf<T>, <T as Config>::MaxJurySize>,
+        pub votes: Vote,
+        pub reason: BoundedVec<u8, <T as Config>::MaxReasonLength>,
+        pub jury: BoundedVec<AccountIdOf<T>, <T as Config>::MaxJurySize>,
     }
 
     #[derive(Encode, Decode, PartialEq, Eq, Clone, Debug, TypeInfo, MaxEncodedLen)]
     pub struct Vote {
-        yay: u32,
-        nay: u32,
+        pub yay: u32,
+        pub nay: u32,
     }
 
-    impl<T: Config> DisputeHooks<T::DisputeKey> for Pallet<T> {
-        // FELIX REVIEW: i think DisputeHooks should be impled where T: pallet_proposals + pallet_refund.
-        // if you impl it for pallet_dispute then how do we know on pallet_proposals that the dispute is complete?
-        // call me pls
-        fn on_dispute_complete(
-            dispute_key: <T as Config>::DisputeKey,
-        ) -> Result<(), DispatchError> {
-            // verifying whether the given dispute exists, if not throwing exception
-            ensure!(
-                !Disputes::<T>::contains_key(dispute_key.clone()),
-                Error::<T>::DisputeDoesNotExist
-            );
-            //SHANKAR: getting dispute so that to be useful to emit via event confirm with FELIX?
-            // FELIX REVIEW, again, only emit events when you know the data is required.
-            let dispute = Disputes::<T>::get(dispute_key.clone());
-            //emitting the dispute as it is completed
-            Self::deposit_event(Event::DisputeCompleted {});
-            //removing the dispute once its being completed
-            Disputes::<T>::remove(dispute_key);
-            //SHANKAR: How about handling the refund(distrition of the fund) logic here, need to discuss with FELIX
-            //Also we need to return outcome type correct here?
-            Ok(())
-        }
-
-        fn on_dispute_cancel(dispute_key: <T as Config>::DisputeKey) -> Result<(), DispatchError> {
-            // verifying whether the given dispute exists, if not throwing exception
-            ensure!(
-                !Disputes::<T>::contains_key(dispute_key.clone()),
-                Error::<T>::DisputeDoesNotExist
-            );
-            //SHANKAR: getting dispute so that to be useful to emit via event confirm with FELIX?
-            let dispute = Disputes::<T>::get(dispute_key.clone());
-            //emitting the dispute as it is cancelled
-            Self::deposit_event(Event::DisputeCancelled {});
-            //removing the dispute once its being cancelled
-            Disputes::<T>::remove(dispute_key);
-            //SHANKAR: Confirming incase of cancellation there is no need for any refund logic right?
-            Ok(())
-        }
-    }
     enum Outcome {
         Refund,
         Continue,
