+++ pallets/proposals/src/lib.rs
@@ -37,6 +37,9 @@ pub use weights::*;
 pub mod migration;
 
 pub mod impls;
+mod benchmarking;
+mod benchmarking;
+
 pub use impls::*;
 pub type ProjectKey = u32;
 pub type MilestoneKey = u32;
